requirements.txt
Flask==3.0.0
pandas==2.1.4
reportlab==4.0.7
requests==2.31.0
python-dotenv 
supabase